# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: ReelRoulette API
  version: 0.7.0
  description: Contract-first API with M6b refresh pipeline, status/settings APIs, and thumbnail artifact retrieval endpoints.
paths:
...
  /api/pair:
    get:
      summary: Pair client and issue auth cookie
      operationId: getPair
      description: |
        Pairing endpoint for auth-required mode.
        Accepts token via query (`?token=...`) and issues the pairing cookie when valid.
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Pairing succeeded or auth disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Pair client and issue auth cookie
      operationId: postPair
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairRequest"
      responses:
        "200":
          description: Pairing succeeded or auth disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/presets:
    get:
      summary: Get available presets
      operationId: getPresets
      responses:
        "200":
          description: Preset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PresetResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Replace preset catalog
      operationId: postPresets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/FilterPresetSnapshot"
      responses:
        "200":
          description: Preset catalog updated
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sources:
    get:
      summary: Get library sources with global enabled states
      operationId: getSources
      responses:
        "200":
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sources/{sourceId}/enabled:
    post:
      summary: Update global enabled state for a source
      operationId: postSourceEnabled
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSourceEnabledRequest"
      responses:
        "200":
          description: Source state updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "404":
          description: Source not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/presets/match:
    post:
      summary: Match filter state to preset
      operationId: postPresetMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresetMatchRequest"
      responses:
        "200":
          description: Preset match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresetMatchResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/random:
    post:
      summary: Request a random media item
      operationId: postRandom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RandomRequest"
      responses:
        "200":
          description: Random media result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RandomResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/favorite:
    post:
      summary: Update favorite state for an item
      operationId: postFavorite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequest"
      responses:
        "200":
          description: Favorite updated
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/blacklist:
    post:
      summary: Update blacklist state for an item
      operationId: postBlacklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlacklistRequest"
      responses:
        "200":
          description: Blacklist updated
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/record-playback:
    post:
      summary: Record playback for an item
      operationId: postRecordPlayback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordPlaybackRequest"
      responses:
        "200":
          description: Playback recorded
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/library-states:
    post:
      summary: Fetch authoritative item state snapshots
      operationId: postLibraryStates
      description: |
        Re-fetch endpoint used during SSE reconnect/resync flows.
        Clients submit known item paths and receive current favorite/blacklist
        state with the latest revision observed for each item.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryStatesRequest"
      responses:
        "200":
          description: State snapshots for requested paths (or all known states when no paths are provided)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryStateResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/model:
    post:
      summary: Get tag editor model snapshot
      operationId: postTagEditorModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagEditorModelRequest"
      responses:
        "200":
          description: Tag editor model snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagEditorModelResponse"
        "401":
          description: Unauthorized when auth is required and request is not paired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/apply-item-tags:
    post:
      summary: Apply batch add/remove tag deltas to items
      operationId: postTagEditorApplyItemTags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyItemTagsRequest"
      responses:
        "200":
          description: Tag delta applied
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/upsert-category:
    post:
      summary: Create or update category metadata
      operationId: postTagEditorUpsertCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCategoryRequest"
      responses:
        "200":
          description: Category upserted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/upsert-tag:
    post:
      summary: Create or update tag metadata
      operationId: postTagEditorUpsertTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTagRequest"
      responses:
        "200":
          description: Tag upserted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/rename-tag:
    post:
      summary: Rename tag and optionally move category
      operationId: postTagEditorRenameTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameTagRequest"
      responses:
        "200":
          description: Tag renamed
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/delete-tag:
    post:
      summary: Delete a tag
      operationId: postTagEditorDeleteTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteTagRequest"
      responses:
        "200":
          description: Tag deleted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/delete-category:
    post:
      summary: Delete a category and optionally reassign tags
      operationId: postTagEditorDeleteCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCategoryRequest"
      responses:
        "200":
          description: Category deleted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/sync-catalog:
    post:
      summary: Replace server tag catalog snapshot from client source-of-truth
      operationId: postTagEditorSyncCatalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncTagCatalogRequest"
      responses:
        "200":
          description: Catalog synchronized
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tag-editor/sync-item-tags:
    post:
      summary: Hydrate item-tag snapshots for requested items
      operationId: postTagEditorSyncItemTags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncItemTagsRequest"
      responses:
        "200":
          description: Item tags synchronized
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/refresh/start:
    post:
      summary: Start unified refresh run
      operationId: postRefreshStart
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshStartRequest"
      responses:
        "200":
          description: Refresh run accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshStartResponse"
        "409":
          description: Refresh already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/refresh/settings:
    get:
      summary: Get core-owned refresh settings
      operationId: getRefreshSettings
      responses:
        "200":
          description: Refresh settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshSettingsSnapshot"
    post:
      summary: Update core-owned refresh settings
      operationId: postRefreshSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshSettingsSnapshot"
      responses:
        "200":
          description: Updated refresh settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshSettingsSnapshot"
  /api/thumbnail/{itemId}:
    get:
      summary: Get generated thumbnail by item id
      operationId: getThumbnailByItemId
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thumbnail image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Thumbnail not found
...
components:
  schemas:
...
    PairRequest:
      type: object
      properties:
        token:
          type: string
          nullable: true
    PairResponse:
      type: object
      required:
        - paired
        - message
      properties:
        paired:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    PresetResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        summary:
          type: string
          nullable: true
        filterState:
          type: object
          nullable: true
          additionalProperties: true
    SourceResponse:
      type: object
      required:
        - id
        - rootPath
        - isEnabled
      properties:
        id:
          type: string
        rootPath:
          type: string
        displayName:
          type: string
          nullable: true
        isEnabled:
          type: boolean
    UpdateSourceEnabledRequest:
      type: object
      required:
        - isEnabled
      properties:
        isEnabled:
          type: boolean
    RandomRequest:
      type: object
      properties:
        presetId:
          type: string
        filterState:
          type: object
          nullable: true
          additionalProperties: true
        clientId:
          type: string
          nullable: true
        includeVideos:
          type: boolean
          default: true
        includePhotos:
          type: boolean
          default: true
        randomizationMode:
          type: string
          nullable: true
...
    PresetMatchRequest:
      type: object
      properties:
        filterState:
          type: object
          nullable: true
          additionalProperties: true
    PresetMatchResponse:
      type: object
      required:
        - matched
      properties:
        matched:
          type: boolean
        presetId:
          type: string
          nullable: true
        presetName:
          type: string
          nullable: true
    FavoriteRequest:
      type: object
      required:
        - path
        - isFavorite
      properties:
        path:
          type: string
        isFavorite:
          type: boolean
    BlacklistRequest:
      type: object
      required:
        - path
        - isBlacklisted
      properties:
        path:
          type: string
        isBlacklisted:
          type: boolean
...
    FilterPresetSnapshot:
      type: object
      required:
        - name
        - filterState
      properties:
        name:
          type: string
        filterState:
          type: object
          additionalProperties: true
...